+++ src/Jackalope/Transport/DoctrineDBAL/Client.php
@@ -2,11 +2,12 @@
 
 namespace Jackalope\Transport\DoctrineDBAL;
 
+use Doctrine\DBAL\ArrayParameterType;
 use Doctrine\DBAL\Connection;
 use Doctrine\DBAL\Driver\PDO\Connection as PDOConnection;
 use Doctrine\DBAL\Exception as DBALException;
 use Doctrine\DBAL\ParameterType;
-use Doctrine\DBAL\Platforms\MySQLPlatform;
+use Doctrine\DBAL\Platforms\AbstractMySQLPlatform;
 use Doctrine\DBAL\Platforms\PostgreSQL94Platform;
 use Doctrine\DBAL\Platforms\PostgreSQLPlatform;
 use Doctrine\DBAL\Platforms\SqlitePlatform;
@@ -256,7 +257,7 @@ class Client extends BaseTransport implements QueryTransport, WritingInterface, 
     /**
      * @throws NotImplementedException
      */
-    public function createWorkspace(string $name, string $srcWorkspace = null): void
+    public function createWorkspace(string $name, ?string $srcWorkspace = null): void
     {
         if (null !== $srcWorkspace) {
             throw new NotImplementedException('Creating workspace as clone of existing workspace not supported');
@@ -319,7 +320,7 @@ class Client extends BaseTransport implements QueryTransport, WritingInterface, 
         }
     }
 
-    public function login(CredentialsInterface $credentials = null, string $workspaceName = null): string
+    public function login(?CredentialsInterface $credentials = null, ?string $workspaceName = null): string
     {
         $this->credentials = $credentials;
 
@@ -363,7 +364,7 @@ class Client extends BaseTransport implements QueryTransport, WritingInterface, 
     }
 
     /**
-     * This will control the collate which is being used on MySQL when querying nodes. It will be autodetected by just
+     * This will control the collation which is being used on MySQL when querying nodes. It will be autodetected by just
      * appending _bin to the current charset, which is good enough in most cases.
      */
     public function setCaseSensitiveEncoding(string $encoding): void
@@ -374,17 +375,20 @@ class Client extends BaseTransport implements QueryTransport, WritingInterface, 
     /**
      * Returns the collate which is being used on MySQL when querying nodes.
      */
-    private function getCaseSensitiveEncoding(): string
+    private function getMySQLCaseSensitiveEncoding(): string
     {
         if (null !== $this->caseSensitiveEncoding) {
             return $this->caseSensitiveEncoding;
         }
 
         $params = $this->conn->getParams();
-        $charset = $params['charset'] ?? 'utf8';
         if (isset($params['defaultTableOptions']['collate'])) {
             return $this->caseSensitiveEncoding = $params['defaultTableOptions']['collate'];
         }
+        if (!array_key_exists('charset', $params)) {
+            throw new \InvalidArgumentException('For MySQL, the Doctrine dbal connection must have either "collate" or "charset" configured. Alternatively, you can set the encoding with '.__CLASS__.'::setCaseSensitiveEncoding().');
+        }
+        $charset = $params['charset'] ?? 'utf8';
 
         return $this->caseSensitiveEncoding = 'binary' === $charset ? $charset : $charset.'_bin';
     }
@@ -543,7 +547,7 @@ class Client extends BaseTransport implements QueryTransport, WritingInterface, 
      */
     private function executeChunkedUpdate(string $query, array $params): void
     {
-        $types = [Connection::PARAM_INT_ARRAY];
+        $types = [ArrayParameterType::INTEGER];
 
         if ($this->getConnection()->getDatabasePlatform() instanceof SqlitePlatform) {
             foreach (array_chunk($params, self::SQLITE_MAXIMUM_IN_PARAM_COUNT) as $chunk) {
@@ -554,7 +558,7 @@ class Client extends BaseTransport implements QueryTransport, WritingInterface, 
         }
     }
 
-    public function copyNode(string $srcAbsPath, string $destAbsPath, string $srcWorkspace = null): void
+    public function copyNode(string $srcAbsPath, string $destAbsPath, ?string $srcWorkspace = null): void
     {
         $this->assertLoggedIn();
 
@@ -921,11 +925,11 @@ class Client extends BaseTransport implements QueryTransport, WritingInterface, 
             if ($this->getConnection()->getDatabasePlatform() instanceof SqlitePlatform) {
                 $missingTargets = [];
                 foreach (array_chunk($params, self::SQLITE_MAXIMUM_IN_PARAM_COUNT) as $chunk) {
-                    $stmt = $this->getConnection()->executeQuery($query, [$chunk], [Connection::PARAM_INT_ARRAY]);
+                    $stmt = $this->getConnection()->executeQuery($query, [$chunk], [ArrayParameterType::INTEGER]);
                     $missingTargets = array_merge($missingTargets, array_column($stmt->fetchAllNumeric(), 0));
                 }
             } else {
-                $stmt = $this->getConnection()->executeQuery($query, [$params], [Connection::PARAM_INT_ARRAY]);
+                $stmt = $this->getConnection()->executeQuery($query, [$params], [ArrayParameterType::INTEGER]);
                 $missingTargets = array_column($stmt->fetchAllNumeric(), 0);
             }
             if ($missingTargets) {
@@ -1257,14 +1261,14 @@ class Client extends BaseTransport implements QueryTransport, WritingInterface, 
                 $childrenRows += $this->getConnection()->fetchAllAssociative(
                     $query,
                     [$chunk, $this->workspaceName],
-                    [Connection::PARAM_STR_ARRAY, null]
+                    [ArrayParameterType::STRING, null]
                 );
             }
         } else {
             $childrenRows = $this->getConnection()->fetchAllAssociative(
                 $query,
                 [$paths, $this->workspaceName],
-                [Connection::PARAM_STR_ARRAY, null]
+                [ArrayParameterType::STRING, null]
             );
         }
 
@@ -1360,7 +1364,7 @@ class Client extends BaseTransport implements QueryTransport, WritingInterface, 
      * @param string      $path          Path of the node
      * @param string|null $workspaceName To overwrite the current workspace
      */
-    private function pathExists(string $path, string $workspaceName = null): bool
+    private function pathExists(string $path, ?string $workspaceName = null): bool
     {
         return (bool) $this->getSystemIdForNode($path, $workspaceName);
     }
@@ -1373,7 +1377,7 @@ class Client extends BaseTransport implements QueryTransport, WritingInterface, 
      *
      * @return bool|string The database id
      */
-    private function getSystemIdForNode(string $identifier, string $workspaceName = null)
+    private function getSystemIdForNode(string $identifier, ?string $workspaceName = null)
     {
         if (null === $workspaceName) {
             $workspaceName = $this->workspaceName;
@@ -1383,8 +1387,8 @@ class Client extends BaseTransport implements QueryTransport, WritingInterface, 
             $query = 'SELECT id FROM phpcr_nodes WHERE identifier = ? AND workspace_name = ?';
         } else {
             $platform = $this->getConnection()->getDatabasePlatform();
-            if ($platform instanceof MySQLPlatform) {
-                $query = 'SELECT id FROM phpcr_nodes WHERE path COLLATE '.$this->getCaseSensitiveEncoding().' = ? AND workspace_name = ?';
+            if ($platform instanceof AbstractMySQLPlatform) {
+                $query = 'SELECT id FROM phpcr_nodes WHERE path COLLATE '.$this->getMySQLCaseSensitiveEncoding().' = ? AND workspace_name = ?';
             } else {
                 $query = 'SELECT id FROM phpcr_nodes WHERE path = ? AND workspace_name = ?';
             }
@@ -1428,14 +1432,14 @@ class Client extends BaseTransport implements QueryTransport, WritingInterface, 
                 $all += $this->getConnection()->fetchAllAssociative(
                     $query,
                     [$this->workspaceName, $chunk],
-                    [ParameterType::STRING, Connection::PARAM_STR_ARRAY]
+                    [ParameterType::STRING, ArrayParameterType::STRING]
                 );
             }
         } else {
             $all = $this->getConnection()->fetchAllAssociative(
                 $query,
                 [$this->workspaceName, $identifiers],
-                [ParameterType::STRING, Connection::PARAM_STR_ARRAY]
+                [ParameterType::STRING, ArrayParameterType::STRING]
             );
         }
 
@@ -1457,7 +1461,7 @@ class Client extends BaseTransport implements QueryTransport, WritingInterface, 
         return $nodes;
     }
 
-    public function getNodePathForIdentifier(string $uuid, string $workspace = null): string
+    public function getNodePathForIdentifier(string $uuid, ?string $workspace = null): string
     {
         if (null !== $workspace) {
             throw new NotImplementedException('Specifying the workspace is not yet supported.');
@@ -1679,10 +1683,20 @@ class Client extends BaseTransport implements QueryTransport, WritingInterface, 
             throw new PathNotFoundException("Parent of the destination path '".$destAbsPath."' has to exist.");
         }
 
-        $query = 'SELECT path, id FROM phpcr_nodes WHERE path LIKE ? OR path = ? AND workspace_name = ? '.$this->getConnection()
-                ->getDatabasePlatform()->getForUpdateSQL()
+        $qb = $this->getConnection()->createQueryBuilder()
+            ->select('path, id')
+            ->from('phpcr_nodes')
+            ->where('(path LIKE :path_prefix OR path = :path) AND workspace_name = :workspace')
         ;
-        $stmt = $this->getConnection()->executeQuery($query, [$srcAbsPath.'/%', $srcAbsPath, $this->workspaceName]);
+        if (!$this->getConnection()->getDatabasePlatform() instanceof SqlitePlatform) {
+            $qb->forUpdate();
+        }
+
+        $stmt = $this->getConnection()->executeQuery($qb->getSQL(), [
+            'path_prefix' => $srcAbsPath.'/%',
+            'path' => $srcAbsPath,
+            'workspace' => $this->workspaceName,
+            ]);
 
         /*
          * TODO: https://github.com/jackalope/jackalope-doctrine-dbal/pull/26/files#L0R1057
@@ -1704,15 +1718,10 @@ class Client extends BaseTransport implements QueryTransport, WritingInterface, 
 
         // TODO: Find a better way to do this
         // Calculate CAST type for CASE statement
-        switch ($this->getConnection()->getDatabasePlatform()->getName()) {
-            case 'pgsql':
-                $intType = 'integer';
-                break;
-            case 'mysql':
-                $intType = 'unsigned';
-                break;
-            default:
-                $intType = 'integer';
+
+        $intType = 'integer';
+        if ($this->getConnection()->getDatabasePlatform() instanceof AbstractMySQLPlatform) {
+            $intType = 'unsigned';
         }
 
         $i = 0;
@@ -2357,7 +2366,7 @@ phpcr_type_childs ON phpcr_type_nodes.node_type_id = phpcr_type_childs.node_type
         }
     }
 
-    public function getReferences(string $path, string $name = null): array
+    public function getReferences(string $path, ?string $name = null): array
     {
         return $this->getNodeReferences($path, $name, false);
     }
@@ -2374,7 +2383,7 @@ phpcr_type_childs ON phpcr_type_nodes.node_type_id = phpcr_type_childs.node_type
      *
      * @return string[] list of paths to nodes that reference $path
      */
-    private function getNodeReferences(string $path, string $name = null, bool $weakReference = false): array
+    private function getNodeReferences(string $path, ?string $name = null, bool $weakReference = false): array
     {
         $targetId = $this->getSystemIdForNode($path);